arena: create library header
The number of lifetimes/types of arenas is a compile time constant and the goal is to provide a generic, user composed, arena data structure.
To solve both these problems the `first[]` list & the `*arena[]` pointer list both need to be defined externally.
It is up to the user of this header library to add a list, `free[]`, of `struct arena`'s initialized to zeros
as well as a list, `*arena[]`, of pointers to `struct arena`s with their values being references to corresponding free[] indices.
The library will handle all further logic, ⸺   using both `free[]` & `*arena[]` ⸺   with the user calling corresponding allocate() and deallocate() functions.